groovy bonus

+++ src/main/java/myapp/D1HourValueProvider.java
@@ -1,8 +1,10 @@
 package myapp;
 
+import java.math.BigDecimal;
+
 public class D1HourValueProvider implements HourValueProvider {
     @Override
-    public int getValue() {
-        return 10;
+    public BigDecimal getValue() {
+        return BigDecimal.valueOf(10);
     }
 }
